requirements.txt
alembic==1.14.0
beautifulsoup4==4.8.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
CityLearn==2.3.0
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.40.0
decorator==5.1.1
Deprecated==1.2.15
docker==7.1.0
doe-xstock==1.1.0
eppy==0.5.63
Farama-Notifications==0.0.4
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
gymnasium==1.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mlflow==2.19.0
mlflow-skinny==2.19.0
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
NREL-PySAM==6.0.0
numpy==1.26.4
onnx==1.17.0
openstudio==3.3.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
pyarrow==18.1.0
pydantic==2.7.1
pytest==8.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot3k==1.0.17
pynvml==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.2.2
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tinynumpy==1.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
